Waveguide resonator (#217)
* feat(innexus): add waveguide string resonator with modal/waveguide switching

Implement digital waveguide string resonator (Karplus-Strong/EKS) as an
alternative to the existing modal resonator bank in Innexus. Users can
switch between Modal and Waveguide resonance types with click-free
equal-power cosine crossfade and energy-aware gain matching.

DSP core (dsp/include/krate/dsp/processors/):
- IResonator: shared pure virtual interface for interchangeable resonators
- WaveguideString: two-segment delay loop with loss filter, dispersion
  allpass cascade (4 sections), DC blocker, soft clipper, energy floor,
  dual energy followers, noise burst excitation with pick-position comb
- ModalResonatorBank: adapted to conform to IResonator interface

Plugin integration (plugins/innexus/):
- Three new parameters: ResonanceType, WaveguideStiffness, WaveguidePickPosition
- Voice engine wiring with per-voice WaveguideString instances
- 1024-sample crossfade with equal-power formula and energy gain matching
- Full state save/load and controller registration

Key features:
- Sub-1-cent pitch accuracy (A2-A5) via exact phase delay compensation
- Stiffness control for inharmonicity shaping (Fletcher's formula)
- Pick position comb filter with circular subtraction for deep spectral nulls
- Energy normalisation for consistent loudness across pitch range (within 3 dB)
- ScatteringJunction/PluckJunction stubs for future bow model (Phase 4)

Also fixes 63 clang-tidy warnings across disrumpo, ruinae, and innexus plugins.

Closes #129


* feat(innexus): add resonance type UI with modal/waveguide view switching

Add ResonanceType option menu to the Resonator Bank section that switches
between Modal Resonator Bank and Waveguide String control views. Modal
shows Mix/Decay/Brightness/Stretch/Scatter; Waveguide shows
Mix/Decay/Brightness/Stiffness/PickPosition. Visibility controlled via
parameter-driven 30ms timer, matching the existing Impact exciter pattern.


---------

+++ plugins/innexus/src/processor/processor_messages.cpp
@@ -83,7 +83,7 @@ void Processor::sendDisplayData(Steinberg::Vst::ProcessData& data)
         static_cast<int>(Krate::DSP::kMaxPartials));
 
     // Zero-initialize, then fill active partials from frame data
-    for (int i = 0; i < static_cast<int>(Krate::DSP::kMaxPartials); ++i)
+    for (size_t i = 0; i < Krate::DSP::kMaxPartials; ++i)
     {
         displayDataBuffer_.partialAmplitudes[i] = 0.0f;
         displayDataBuffer_.partialActive[i] = 0;
